refactor: consolidate test directory structure and align docs to reality
- Move tests/integration/ into tests/4_cross_domain/integration/ to
  unify under the numbered domain scheme
- Rewrite tests/README.md removing false "188 files" claim, documenting
  actual structure (13 unit + 12 integration/E2E)
- Add deno.json test:unit, test:integration, test:e2e tasks
- Document Deno.env.set save-set-restore exception in tests/CLAUDE.md
