+++ tests/4_cross_domain/integration/directive_layer_types/fixtures/configs/pattern-matching/pattern-test-app.yml
