+++ tests/4_cross_domain/integration/directive_layer_types/fixtures/prompts/summary/issue/f_task.md
