+++ tests/4_cross_domain/integration/directive_layer_types/fixtures/prompts/to/project/f_issue.md
