+++ tests/4_cross_domain/integration/directive_layer_types/fixtures/schemas/summary/issue.json
