+++ tests/4_cross_domain/integration/directive_layer_types/test_scenarios/path_resolution.md
