+++ tests/CLAUDE.md
@@ -1,6 +1,7 @@
 # 禁止事項
 - プロジェクト直下の examples/ 配下をテストすることは禁ずる。（テスト対象にしない。） examples/ 配下は、テストファイルでも実装ファイルでもない。
 - テストファイル内で、`Deno.env.set` を使ってモードをセットすること
+  - Exception: E2E tests may use save-set-restore pattern (`const original = Deno.env.get(); ... Deno.env.set(original)` in finally block)
 - テストコードでは絶対パス（/Users/username/...）を使用禁止し、相対パス（./、../）またはDeno.cwd()ベースの動的パス解決を必須とする。
 
 # デバッグ出力戦略
